feat: Add game royalty count flag
Introduce a new flag `countInRoyalty` to the `Game` type. This flag controls whether a game's results contribute to the season's Trivia Royalty leaderboard.

When editing or creating games, this flag will be available to toggle. For legacy data, existing games will default to counting towards royalty unless explicitly set otherwise or if they are not 'Regular' type games. The `calculateSeasonStandings` service has been updated to respect this new flag.